refactor: 일일 최대값 batch processor의 반환 dto의 필드 타입을 ShortTermWeatherForecast 타입에 맞게 수정

+++ batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/daily_high/processor/dto/DailyHighProcessorResponseDto.java
@@ -4,7 +4,7 @@ public record DailyHighProcessorResponseDto(
         int nx,
         int ny,
         double maxPrecipitation,
-        double maxHumidity,
+        int maxHumidity,
         double maxWindSpeed,
         int windDirection,
         int maxTemperature
